Add email template to certificate module and send function
A new file named email.tsx is added which contains the template for emails to be sent via the certificate module. In addition, the package.json file is updated with necessary dependencies required for handling emails. Also, the index.ts file is expanded with a function to send emails with attachments using nodemailer and react email renderer. A sqlite database is used to keep track of sent emails.